Allow `colspan`/`rowspan` on table cells
Because I have a module, that has a complex table in it...

By the way, any reason the styling of tables on npm site is the way it is? Github centers (or, rather, does not reset to `left`) the `th`. Also, borders do help... https://www.npmjs.com/package/gampee#params vs https://github.com/insidewarehouse/gampee#params