Convert emoji to shortcode names in generated heading IDs
This commit changes the headings.js module into a markdown-it plugin
that does the same slug generation and HTML modifications the older code
did, but in such a way that makes it possible to generate heading
anchors that match what GitHub generates.

We do this by creating a markdown-it renderer rule for 'heading_open'
tokens, and performing a few different operations:

1. Check to see if the heading already contains a link. If so, skip the
   heading.
2. Render to HTML a version of the heading's markdown that has had
   github-style `:emoji:` gemoji strings held back from being converted
   to unicode (the markdown-it-emoji plugin does its translation when
   the markdown is parsed, not when it's rendered). The goal here is to
   let github-slugger properly drop unicode emoji from generated slugs,
   but include the shortcode names for gemoji.
3. Build the slug with github-slugger. We do this by running it through
   `cheerio` first so we can match our previous behavior, i.e., deal
   with embedded HTML more cleanly.
4. Add the `deep-link` class to the heading.
5. Wrap the contents of the heading in a link using the generated slug
   in its `href` attribute.

This commit also includes a handful of tests that put the slug
generation code through its paces, by trying various combinations of
unicode emoji, gemoji, hyphens, colons, etc...

And finally, we trim our node_modules a little by only depending on the
specific `lodash` modules we're actually using, rather than all of
lodash itself.