Dont require 'highlights' in browserify environment.
This change should make it possible to use marky-markdown with
browserify as long as the {highlightSyntax: false} option is used.
Using with {highlightSyntax: true} would generate a runtime error
in the browser.

Example command:
  browserify main.js -i highlights > bundle.js

Also note that by not including the highlights library, the file
size of the resulting bundle is reduced by half.