Allow <h1 align="center"> like GitHub
This adds support for center-aligned H1 tags when building a centered heading for a README, which is possible at GitHub, but comes out weird at NPM. Particularly when repo shields/badges are centered, or a header image appears right underneath it. These are things that _are_ centered at NPM, while the heading itself is not — this fixes that.