feature: add rel=nofollow to links if the `nofollow` option is set
Added a plugin that adds rel=nofollow to links and does no more.
Added an option to control if the plugin is loaded, defaulted to true,
and documented the option.
Wrote tests to verify that the rel attr is added, and that the option is
respected.
Added `rel` to the allowed list of attributes in the sanitizer.